feat(backbones): reuse pretrained HuggingFace encoders as backbones (#680)
Reuse an external pretrained image backbone from HuggingFace `transformers`
(ConvNeXtV2, ResNet, Swinv2, DINOv2, ...) as the backbone of any sleap-nn model —
frozen or fine-tuned — for pose, centroid, or segmentation heads. Rides the
existing Model/get_backbone path via a new `pretrained` member on
`backbone_config`; no new model class, no head changes.

- Two auto-selected modes: `decoder` (hierarchical pyramid -> sleap-nn Decoder for
  spatial heads) and `encoder` (isotropic-ViT pooled bottleneck for pooled heads).
- Wrapper forces float32, pins an optional `revision`, bakes in AutoImageProcessor
  mean/std, probes strides, re-applies pretrained weights after xavier init, and
  supports `freeze` (optimizer filters to requires_grad params).
- `transformers` is a lazy `backbones` optional extra, so the default install is
  unchanged.
- ONNX exporter gains an opt-in torch-vs-onnxruntime numerical-parity check for
  transformer backbones.
- Docs: a Pretrained Backbones guide (supported-model table, modes, freeze/
  fine-tune, gating/HF_TOKEN/offline workflow), a model-config section, a sample
  config, and mkdocs nav. Network-free wrapper/config tests.

Closes #680.

+++ sleap_nn/architectures/encoder_decoder.py
@@ -599,7 +599,7 @@ class Decoder(nn.Module):
         up_blocks: int = 4,
         down_blocks: int = 3,
         stem_blocks: int = 0,
-        filters_rate: int = 2,
+        filters_rate: float = 2,
         convs_per_block: int = 2,
         kernel_size: int = 3,
         block_contraction: bool = False,
